+++ oshi-core-java25/src/main/java/oshi/ffm/mac/IOKitFunctions.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2025 The OSHI Project Contributors
+ * Copyright 2025-2026 The OSHI Project Contributors
  * SPDX-License-Identifier: MIT
  */
 package oshi.ffm.mac;
@@ -8,9 +8,11 @@ import static java.lang.foreign.ValueLayout.ADDRESS;
 import static java.lang.foreign.ValueLayout.JAVA_BOOLEAN;
 import static java.lang.foreign.ValueLayout.JAVA_DOUBLE;
 import static java.lang.foreign.ValueLayout.JAVA_INT;
+import static java.lang.foreign.ValueLayout.JAVA_LONG;
 
 import java.lang.foreign.FunctionDescriptor;
 import java.lang.foreign.MemorySegment;
+import java.lang.foreign.SymbolLookup;
 import java.lang.invoke.MethodHandle;
 
 import oshi.ffm.ForeignFunctions;
@@ -20,13 +22,15 @@ public final class IOKitFunctions extends ForeignFunctions {
     private IOKitFunctions() {
     }
 
+    private static final SymbolLookup IOKIT_LIBRARY = frameworkLookup("IOKit");
+
     /*
      * IOKitLib.h
      */
 
     // kern_return_t IOMasterPort(mach_port_t bootstrapPort, mach_port_t *mainPort);
 
-    private static final MethodHandle IOMasterPort = LINKER.downcallHandle(SYMBOL_LOOKUP.findOrThrow("IOMasterPort"),
+    private static final MethodHandle IOMasterPort = LINKER.downcallHandle(IOKIT_LIBRARY.findOrThrow("IOMasterPort"),
             FunctionDescriptor.of(JAVA_INT, JAVA_INT, ADDRESS));
 
     public static int IOMasterPort(int bootstrapPort, MemorySegment port) throws Throwable {
@@ -36,7 +40,7 @@ public final class IOKitFunctions extends ForeignFunctions {
     // io_registry_entry_t IORegistryGetRootEntry(mach_port_t mainPort);
 
     private static final MethodHandle IORegistryGetRootEntry = LINKER.downcallHandle(
-            SYMBOL_LOOKUP.findOrThrow("IORegistryGetRootEntry"), FunctionDescriptor.of(ADDRESS, JAVA_INT));
+            IOKIT_LIBRARY.findOrThrow("IORegistryGetRootEntry"), FunctionDescriptor.of(ADDRESS, JAVA_INT));
 
     public static MemorySegment IORegistryGetRootEntry(int masterPort) throws Throwable {
         return (MemorySegment) IORegistryGetRootEntry.invokeExact(masterPort);
@@ -45,7 +49,7 @@ public final class IOKitFunctions extends ForeignFunctions {
     // CFMutableDictionaryRef IOServiceNameMatching(const char *name);
 
     private static final MethodHandle IOServiceNameMatching = LINKER.downcallHandle(
-            SYMBOL_LOOKUP.findOrThrow("IOServiceNameMatching"), FunctionDescriptor.of(ADDRESS, ADDRESS));
+            IOKIT_LIBRARY.findOrThrow("IOServiceNameMatching"), FunctionDescriptor.of(ADDRESS, ADDRESS));
 
     public static MemorySegment IOServiceNameMatching(MemorySegment name) throws Throwable {
         return (MemorySegment) IOServiceNameMatching.invokeExact(name);
@@ -54,7 +58,7 @@ public final class IOKitFunctions extends ForeignFunctions {
     // CFMutableDictionaryRef IOServiceMatching(const char *name);
 
     private static final MethodHandle IOServiceMatching = LINKER
-            .downcallHandle(SYMBOL_LOOKUP.findOrThrow("IOServiceMatching"), FunctionDescriptor.of(ADDRESS, ADDRESS));
+            .downcallHandle(IOKIT_LIBRARY.findOrThrow("IOServiceMatching"), FunctionDescriptor.of(ADDRESS, ADDRESS));
 
     public static MemorySegment IOServiceMatching(MemorySegment name) throws Throwable {
         return (MemorySegment) IOServiceMatching.invokeExact(name);
@@ -63,7 +67,7 @@ public final class IOKitFunctions extends ForeignFunctions {
     // io_service_t IOServiceGetMatchingService(mach_port_t mainPort, CFDictionaryRef matching);
 
     private static final MethodHandle IOServiceGetMatchingService = LINKER.downcallHandle(
-            SYMBOL_LOOKUP.findOrThrow("IOServiceGetMatchingService"),
+            IOKIT_LIBRARY.findOrThrow("IOServiceGetMatchingService"),
             FunctionDescriptor.of(ADDRESS, JAVA_INT, ADDRESS));
 
     public static MemorySegment IOServiceGetMatchingService(int masterPort, MemorySegment matchingDict)
@@ -75,7 +79,7 @@ public final class IOKitFunctions extends ForeignFunctions {
     // *existing);
 
     private static final MethodHandle IOServiceGetMatchingServices = LINKER.downcallHandle(
-            SYMBOL_LOOKUP.findOrThrow("IOServiceGetMatchingServices"),
+            IOKIT_LIBRARY.findOrThrow("IOServiceGetMatchingServices"),
             FunctionDescriptor.of(JAVA_INT, JAVA_INT, ADDRESS, ADDRESS));
 
     public static int IOServiceGetMatchingServices(int masterPort, MemorySegment matchingDict, MemorySegment iterator)
@@ -86,7 +90,7 @@ public final class IOKitFunctions extends ForeignFunctions {
     // CFMutableDictionaryRef IOBSDNameMatching(mach_port_t mainPort, uint32_t options, const char *bsdName);
 
     private static final MethodHandle IOBSDNameMatching = LINKER.downcallHandle(
-            SYMBOL_LOOKUP.findOrThrow("IOBSDNameMatching"),
+            IOKIT_LIBRARY.findOrThrow("IOBSDNameMatching"),
             FunctionDescriptor.of(ADDRESS, JAVA_INT, JAVA_INT, ADDRESS));
 
     public static MemorySegment IOBSDNameMatching(int masterPort, int options, MemorySegment bsdName) throws Throwable {
@@ -96,7 +100,7 @@ public final class IOKitFunctions extends ForeignFunctions {
     // kern_return_t IOObjectRelease(io_object_t object);
 
     private static final MethodHandle IOObjectRelease = LINKER
-            .downcallHandle(SYMBOL_LOOKUP.findOrThrow("IOObjectRelease"), FunctionDescriptor.of(JAVA_INT, ADDRESS));
+            .downcallHandle(IOKIT_LIBRARY.findOrThrow("IOObjectRelease"), FunctionDescriptor.of(JAVA_INT, ADDRESS));
 
     public static int IOObjectRelease(MemorySegment object) throws Throwable {
         return (int) IOObjectRelease.invokeExact(object);
@@ -105,7 +109,7 @@ public final class IOKitFunctions extends ForeignFunctions {
     // boolean_t IOObjectConformsTo(io_object_t object, const io_name_t className);
 
     private static final MethodHandle IOObjectConformsTo = LINKER.downcallHandle(
-            SYMBOL_LOOKUP.findOrThrow("IOObjectConformsTo"), FunctionDescriptor.of(JAVA_BOOLEAN, ADDRESS, ADDRESS));
+            IOKIT_LIBRARY.findOrThrow("IOObjectConformsTo"), FunctionDescriptor.of(JAVA_BOOLEAN, ADDRESS, ADDRESS));
 
     public static boolean IOObjectConformsTo(MemorySegment object, MemorySegment className) throws Throwable {
         return (boolean) IOObjectConformsTo.invokeExact(object, className);
@@ -114,7 +118,7 @@ public final class IOKitFunctions extends ForeignFunctions {
     // io_object_t IOIteratorNext(io_iterator_t iterator);
 
     private static final MethodHandle IOIteratorNext = LINKER
-            .downcallHandle(SYMBOL_LOOKUP.findOrThrow("IOIteratorNext"), FunctionDescriptor.of(ADDRESS, ADDRESS));
+            .downcallHandle(IOKIT_LIBRARY.findOrThrow("IOIteratorNext"), FunctionDescriptor.of(ADDRESS, ADDRESS));
 
     public static MemorySegment IOIteratorNext(MemorySegment iterator) throws Throwable {
         return (MemorySegment) IOIteratorNext.invokeExact(iterator);
@@ -123,7 +127,7 @@ public final class IOKitFunctions extends ForeignFunctions {
     // kern_return_t IORegistryEntryGetRegistryEntryID(io_registry_entry_t entry, uint64_t *entryID);
 
     private static final MethodHandle IORegistryEntryGetRegistryEntryID = LINKER.downcallHandle(
-            SYMBOL_LOOKUP.findOrThrow("IORegistryEntryGetRegistryEntryID"),
+            IOKIT_LIBRARY.findOrThrow("IORegistryEntryGetRegistryEntryID"),
             FunctionDescriptor.of(JAVA_INT, ADDRESS, ADDRESS));
 
     public static int IORegistryEntryGetRegistryEntryID(MemorySegment entry, MemorySegment id) throws Throwable {
@@ -133,7 +137,7 @@ public final class IOKitFunctions extends ForeignFunctions {
     // kern_return_t IORegistryEntryGetName(io_registry_entry_t entry, io_name_t name);
 
     private static final MethodHandle IORegistryEntryGetName = LINKER.downcallHandle(
-            SYMBOL_LOOKUP.findOrThrow("IORegistryEntryGetName"), FunctionDescriptor.of(JAVA_INT, ADDRESS, ADDRESS));
+            IOKIT_LIBRARY.findOrThrow("IORegistryEntryGetName"), FunctionDescriptor.of(JAVA_INT, ADDRESS, ADDRESS));
 
     public static int IORegistryEntryGetName(MemorySegment entry, MemorySegment name) throws Throwable {
         return (int) IORegistryEntryGetName.invokeExact(entry, name);
@@ -143,7 +147,7 @@ public final class IOKitFunctions extends ForeignFunctions {
     // *iterator);
 
     private static final MethodHandle IORegistryEntryGetChildIterator = LINKER.downcallHandle(
-            SYMBOL_LOOKUP.findOrThrow("IORegistryEntryGetChildIterator"),
+            IOKIT_LIBRARY.findOrThrow("IORegistryEntryGetChildIterator"),
             FunctionDescriptor.of(JAVA_INT, ADDRESS, ADDRESS, ADDRESS));
 
     public static int IORegistryEntryGetChildIterator(MemorySegment entry, MemorySegment plane, MemorySegment iter)
@@ -155,7 +159,7 @@ public final class IOKitFunctions extends ForeignFunctions {
     // *child);
 
     private static final MethodHandle IORegistryEntryGetChildEntry = LINKER.downcallHandle(
-            SYMBOL_LOOKUP.findOrThrow("IORegistryEntryGetChildEntry"),
+            IOKIT_LIBRARY.findOrThrow("IORegistryEntryGetChildEntry"),
             FunctionDescriptor.of(JAVA_INT, ADDRESS, ADDRESS, ADDRESS));
 
     public static int IORegistryEntryGetChildEntry(MemorySegment entry, MemorySegment plane, MemorySegment child)
@@ -167,7 +171,7 @@ public final class IOKitFunctions extends ForeignFunctions {
     // *parent);
 
     private static final MethodHandle IORegistryEntryGetParentEntry = LINKER.downcallHandle(
-            SYMBOL_LOOKUP.findOrThrow("IORegistryEntryGetParentEntry"),
+            IOKIT_LIBRARY.findOrThrow("IORegistryEntryGetParentEntry"),
             FunctionDescriptor.of(JAVA_INT, ADDRESS, ADDRESS, ADDRESS));
 
     public static int IORegistryEntryGetParentEntry(MemorySegment entry, MemorySegment plane, MemorySegment parent)
@@ -180,7 +184,7 @@ public final class IOKitFunctions extends ForeignFunctions {
     // options);
 
     private static final MethodHandle IORegistryEntryCreateCFProperty = LINKER.downcallHandle(
-            SYMBOL_LOOKUP.findOrThrow("IORegistryEntryCreateCFProperty"),
+            IOKIT_LIBRARY.findOrThrow("IORegistryEntryCreateCFProperty"),
             FunctionDescriptor.of(ADDRESS, ADDRESS, ADDRESS, ADDRESS, JAVA_INT));
 
     public static MemorySegment IORegistryEntryCreateCFProperty(MemorySegment entry, MemorySegment key,
@@ -193,7 +197,7 @@ public final class IOKitFunctions extends ForeignFunctions {
     // allocator, IOOptionBits options);
 
     private static final MethodHandle IORegistryEntryCreateCFProperties = LINKER.downcallHandle(
-            SYMBOL_LOOKUP.findOrThrow("IORegistryEntryCreateCFProperties"),
+            IOKIT_LIBRARY.findOrThrow("IORegistryEntryCreateCFProperties"),
             FunctionDescriptor.of(JAVA_INT, ADDRESS, ADDRESS, ADDRESS, JAVA_INT));
 
     public static int IORegistryEntryCreateCFProperties(MemorySegment entry, MemorySegment properties,
@@ -206,7 +210,7 @@ public final class IOKitFunctions extends ForeignFunctions {
     // allocator, IOOptionBits options);
 
     private static final MethodHandle IORegistryEntrySearchCFProperty = LINKER.downcallHandle(
-            SYMBOL_LOOKUP.findOrThrow("IORegistryEntrySearchCFProperty"),
+            IOKIT_LIBRARY.findOrThrow("IORegistryEntrySearchCFProperty"),
             FunctionDescriptor.of(ADDRESS, ADDRESS, ADDRESS, ADDRESS, ADDRESS, JAVA_INT));
 
     public static MemorySegment IORegistryEntrySearchCFProperty(MemorySegment entry, MemorySegment plane,
@@ -221,7 +225,7 @@ public final class IOKitFunctions extends ForeignFunctions {
     // CFTypeRef IOPSCopyPowerSourcesInfo(void);
 
     private static final MethodHandle IOPSCopyPowerSourcesInfo = LINKER
-            .downcallHandle(SYMBOL_LOOKUP.findOrThrow("IOPSCopyPowerSourcesInfo"), FunctionDescriptor.of(ADDRESS));
+            .downcallHandle(IOKIT_LIBRARY.findOrThrow("IOPSCopyPowerSourcesInfo"), FunctionDescriptor.of(ADDRESS));
 
     public static MemorySegment IOPSCopyPowerSourcesInfo() throws Throwable {
         return (MemorySegment) IOPSCopyPowerSourcesInfo.invokeExact();
@@ -230,7 +234,7 @@ public final class IOKitFunctions extends ForeignFunctions {
     // CFArrayRef IOPSCopyPowerSourcesList(CFTypeRef blob);
 
     private static final MethodHandle IOPSCopyPowerSourcesList = LINKER.downcallHandle(
-            SYMBOL_LOOKUP.findOrThrow("IOPSCopyPowerSourcesList"), FunctionDescriptor.of(ADDRESS, ADDRESS));
+            IOKIT_LIBRARY.findOrThrow("IOPSCopyPowerSourcesList"), FunctionDescriptor.of(ADDRESS, ADDRESS));
 
     public static MemorySegment IOPSCopyPowerSourcesList(MemorySegment blob) throws Throwable {
         return (MemorySegment) IOPSCopyPowerSourcesList.invokeExact(blob);
@@ -239,7 +243,7 @@ public final class IOKitFunctions extends ForeignFunctions {
     // CFDictionaryRef IOPSGetPowerSourceDescription(CFTypeRef blob, CFTypeRef ps);
 
     private static final MethodHandle IOPSGetPowerSourceDescription = LINKER.downcallHandle(
-            SYMBOL_LOOKUP.findOrThrow("IOPSGetPowerSourceDescription"),
+            IOKIT_LIBRARY.findOrThrow("IOPSGetPowerSourceDescription"),
             FunctionDescriptor.of(ADDRESS, ADDRESS, ADDRESS));
 
     public static MemorySegment IOPSGetPowerSourceDescription(MemorySegment blob, MemorySegment ps) throws Throwable {
@@ -249,10 +253,48 @@ public final class IOKitFunctions extends ForeignFunctions {
     // CFTimeInterval IOPSGetTimeRemainingEstimate(void);
 
     private static final MethodHandle IOPSGetTimeRemainingEstimate = LINKER.downcallHandle(
-            SYMBOL_LOOKUP.findOrThrow("IOPSGetTimeRemainingEstimate"), FunctionDescriptor.of(JAVA_DOUBLE));
+            IOKIT_LIBRARY.findOrThrow("IOPSGetTimeRemainingEstimate"), FunctionDescriptor.of(JAVA_DOUBLE));
 
     public static double IOPSGetTimeRemainingEstimate() throws Throwable {
         return (double) IOPSGetTimeRemainingEstimate.invokeExact();
+    }
+
+    /*
+     * IOKitLib.h - IOConnect / IOService open/close
+     */
+
+    // kern_return_t IOServiceOpen(io_service_t service, task_port_t owningTask, uint32_t type,
+    // io_connect_t *connect);
+
+    private static final MethodHandle IOServiceOpen = LINKER.downcallHandle(IOKIT_LIBRARY.findOrThrow("IOServiceOpen"),
+            FunctionDescriptor.of(JAVA_INT, ADDRESS, JAVA_INT, JAVA_INT, ADDRESS));
+
+    public static int IOServiceOpen(MemorySegment service, int owningTask, int type, MemorySegment connect)
+            throws Throwable {
+        return (int) IOServiceOpen.invokeExact(service, owningTask, type, connect);
+    }
+
+    // kern_return_t IOServiceClose(io_connect_t connect);
+
+    private static final MethodHandle IOServiceClose = LINKER
+            .downcallHandle(IOKIT_LIBRARY.findOrThrow("IOServiceClose"), FunctionDescriptor.of(JAVA_INT, JAVA_INT));
+
+    public static int IOServiceClose(int connect) throws Throwable {
+        return (int) IOServiceClose.invokeExact(connect);
+    }
+
+    // kern_return_t IOConnectCallStructMethod(io_connect_t connection, uint32_t selector,
+    // const void *inputStruct, size_t inputStructCnt,
+    // void *outputStruct, size_t *outputStructCnt);
+
+    private static final MethodHandle IOConnectCallStructMethod = LINKER.downcallHandle(
+            IOKIT_LIBRARY.findOrThrow("IOConnectCallStructMethod"),
+            FunctionDescriptor.of(JAVA_INT, JAVA_INT, JAVA_INT, ADDRESS, JAVA_LONG, ADDRESS, ADDRESS));
+
+    public static int IOConnectCallStructMethod(int connection, int selector, MemorySegment inputStruct,
+            long inputStructCnt, MemorySegment outputStruct, MemorySegment outputStructCnt) throws Throwable {
+        return (int) IOConnectCallStructMethod.invokeExact(connection, selector, inputStruct, inputStructCnt,
+                outputStruct, outputStructCnt);
     }
 
 }