+++ oshi-core-java25/src/main/java/oshi/hardware/platform/mac/MacHardwareAbstractionLayerFFM.java
@@ -1,11 +1,13 @@
 /*
- * Copyright 2025 The OSHI Project Contributors
+ * Copyright 2025-2026 The OSHI Project Contributors
  * SPDX-License-Identifier: MIT
  */
 package oshi.hardware.platform.mac;
 
 import oshi.annotation.concurrent.ThreadSafe;
+import oshi.hardware.ComputerSystem;
 import oshi.hardware.GlobalMemory;
+import oshi.hardware.Sensors;
 
 @ThreadSafe
 public final class MacHardwareAbstractionLayerFFM extends MacHardwareAbstractionLayer {
@@ -14,4 +16,13 @@ public final class MacHardwareAbstractionLayerFFM extends MacHardwareAbstraction
         return new MacGlobalMemoryFFM();
     }
 
+    @Override
+    public ComputerSystem createComputerSystem() {
+        return new MacComputerSystemFFM();
+    }
+
+    @Override
+    public Sensors createSensors() {
+        return new MacSensorsFFM();
+    }
 }