+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 * [#3087](https://github.com/oshi/oshi/pull/3087): Correct PowerSupply units on Linux - [@benplotke](https://github.com/benplotke).
 * [#3014](https://github.com/oshi/oshi/pull/3014): Implement COM, WMI, PDH and FileSystem in Windows FFM - [@rohan-coder02](https://github.com/rohan-coder02).
 * [#3104](https://github.com/oshi/oshi/pull/3104): Migrate remaining mac OperatingSystem JNA to FFM - [@dbwiddis](https://github.com/dbwiddis).
+* [#3109](https://github.com/oshi/oshi/pull/3109): Migrate macOS IOKit/SMC hardware to FFM - [@dbwiddis](https://github.com/dbwiddis).
 
 # 6.10.0 (2026-02-22)